# openapi.yaml
openapi: 3.1.0
info:
  title: JisrVOC API
  version: 1.0.0
  description: >
    Backend contract for the Jisr Voice of Customer Intelligence Platform.
    All endpoints require an authenticated session (Jisr SSO / OIDC) and are
    subject to RBAC. List endpoints are cursor-paginated. This contract is the
    shared interface between the React/TypeScript frontend and the FastAPI
    backend; generate the frontend client from this file.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /overview/volume-trend:
    get:
      tags: [Overview]
      summary: Feedback volume over the last N weeks
      parameters:
        - { name: weeks, in: query, schema: { type: integer, default: 12 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrendPoint" }
...
  /overview/top-themes:
    get:
      tags: [Overview]
      summary: Top themes this week with trend signals
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 5 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ThemeSummary" }
  /feedback:
    get:
      tags: [Feedback]
      summary: Filterable, paginated feedback feed (post-decomposition)
      parameters:
        - { name: source, in: query, schema: { $ref: "#/components/schemas/SourceType" } }
        - { name: area, in: query, schema: { $ref: "#/components/schemas/ProductArea" } }
        - { name: category, in: query, schema: { $ref: "#/components/schemas/Category" } }
        - { name: sentiment, in: query, schema: { $ref: "#/components/schemas/Sentiment" } }
        - { name: urgency, in: query, schema: { $ref: "#/components/schemas/Urgency" } }
        - { name: language, in: query, schema: { $ref: "#/components/schemas/Language" } }
        - { name: segment, in: query, schema: { $ref: "#/components/schemas/Segment" } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: q, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackPage" }
  /feedback/{id}:
    get:
      tags: [Feedback]
      summary: Feedback item detail (raw text + AI tags)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackDetail" }
        "404": { description: Not found }
  /feedback/{id}/tags:
    patch:
      tags: [Feedback]
      summary: Correct AI-generated tags (PM only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagCorrection" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackItem" }
...
  /bets:
    get:
      tags: [Bets]
      summary: Bets grouped for the lifecycle board
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BetStatus" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BetSummary" }
    post:
      tags: [Bets]
      summary: Create a bet (usually promoted from a generated draft)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BetCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BetDetail" }
  /bets/{id}:
    get:
      tags: [Bets]
      summary: Bet detail with evidence list
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BetDetail" }
    patch:
      tags: [Bets]
      summary: Update a bet; a status change triggers HubSpot loop-closure write-back
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BetUpdate" }
      responses:
        "200":
          description: OK; includes write-back result when status changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BetUpdateResult" }
  /customers:
    get:
      tags: [Customers]
      summary: Search companies
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
  /customers/{id}/feedback:
    get:
      tags: [Customers]
      summary: All feedback for a company across sources
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeedbackItem" }
  /customers/{id}/bets:
    get:
      tags: [Customers]
      summary: Bets linked to a company's feedback
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BetSummary" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SourceType: { type: string, enum: [hubspot, zendesk, canny, jira] }
    Category:    { type: string, enum: [pain_point, feature_request, bug_report, how_to_question, praise] }
    ProductArea: { type: string, enum: [core_hr, payroll, jisrpay, onboarding, offboarding, contracts, mobile, integrations, other] }
    Sentiment:   { type: string, enum: [positive, neutral, negative, mixed] }
    Urgency:     { type: string, enum: [low, medium, high] }
    Language:    { type: string, enum: [ar, en, mixed] }
    Segment:     { type: string, enum: [smb, mid_market, enterprise, government] }
    Trend:       { type: string, enum: [new, rising, stable, declining] }
    BetStatus:   { type: string, enum: [draft, in_backlog, in_discovery, in_build, shipped, declined] }
...
    TrendPoint:
      type: object
      properties:
        week_start: { type: string, format: date }
        count: { type: integer }
...
    FeedbackItem:
      type: object
      required: [id, summary_en, source, language]
      properties:
        id: { type: string, format: uuid }
        summary_en: { type: string }
        source: { $ref: "#/components/schemas/SourceType" }
        category: { $ref: "#/components/schemas/Category" }
        area: { $ref: "#/components/schemas/ProductArea" }
        sentiment: { $ref: "#/components/schemas/Sentiment" }
        urgency: { $ref: "#/components/schemas/Urgency" }
        language: { $ref: "#/components/schemas/Language" }
        segment: { $ref: "#/components/schemas/Segment" }
        customer_id: { type: string }
        customer_name: { type: string }
        is_split: { type: boolean }
        parent_ticket_id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
    FeedbackDetail:
      allOf:
        - $ref: "#/components/schemas/FeedbackItem"
        - type: object
          properties:
            raw_text: { type: string, description: "Original source text, preserved verbatim (may be RTL Arabic)" }
            raw_language: { $ref: "#/components/schemas/Language" }
            enrichment_model: { type: string }
            enrichment_confidence: { type: number }
            pm_corrected: { type: boolean }
    FeedbackPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedbackItem" }
        next_cursor: { type: string, nullable: true }
        total: { type: integer }
    TagCorrection:
      type: object
      properties:
        category: { $ref: "#/components/schemas/Category" }
        area: { $ref: "#/components/schemas/ProductArea" }
        sentiment: { $ref: "#/components/schemas/Sentiment" }
        urgency: { $ref: "#/components/schemas/Urgency" }
...
    ThemeSummary:
      type: object
      required: [id, name_en, trend, item_count]
      properties:
        id: { type: string, format: uuid }
        name_en: { type: string }
        description_en: { type: string }
        trend: { $ref: "#/components/schemas/Trend" }
        item_count: { type: integer }
        customer_count: { type: integer }
        vote_weight: { type: integer }
        top_segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
...
    BetSummary:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: "#/components/schemas/BetStatus" }
        problem_snippet: { type: string }
        affected_segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
        est_customer_count: { type: integer }
        why_now: { type: string }
        evidence_count: { type: integer }
        theme_id: { type: string, format: uuid }
    BetDetail:
      allOf:
        - $ref: "#/components/schemas/BetSummary"
        - type: object
          properties:
            problem_statement: { type: string }
            owner_pm: { type: string }
            declined_reason: { type: string }
            evidence:
              type: array
              items: { $ref: "#/components/schemas/FeedbackItem" }
    BetCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        theme_id: { type: string, format: uuid }
        problem_statement: { type: string }
        affected_segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
        est_customer_count: { type: integer }
    BetUpdate:
      type: object
      properties:
        title: { type: string }
        problem_statement: { type: string }
        status: { $ref: "#/components/schemas/BetStatus" }
        declined_reason: { type: string }
    BetUpdateResult:
      allOf:
        - $ref: "#/components/schemas/BetDetail"
        - type: object
          properties:
            writeback:
              type: object
              description: Present when a status change triggered loop-closure
              properties:
                tickets_updated: { type: integer }
                action: { type: string, enum: [note, property_update] }
                status_value: { $ref: "#/components/schemas/BetStatus" }
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        domain: { type: string }
        segment: { $ref: "#/components/schemas/Segment" }
        lifecycle_stage: { type: string }
        industry: { type: string }
        is_prospect: { type: boolean }
...